+++ src/media_enrichment/__init__.py
@@ -1,7 +1,7 @@
 """media-enrichment Skill package.
 
-Version: 0.1.0-dev18
+Version: 0.1.0-dev19
 """
 
-__version__ = "0.1.0-dev18"
+__version__ = "0.1.0-dev19"
 SKILL_NAME = "media-enrichment"